+++ src/PokeAByte.Domain/Models/AppSettings.cs
@@ -1,5 +1,8 @@
-﻿using Microsoft.Extensions.Configuration;
+﻿using System.Text.Json;
+using System.Text.Json.Serialization;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
+using PokeAByte.Domain.Models.Mappers;
 
 namespace PokeAByte.Domain.Models;
 
@@ -16,16 +19,23 @@ static class AppSettingsHelper
 
 public class AppSettings
 {
+    public AppSettings()
+    {
+        RETROARCH_LISTEN_IP_ADDRESS = "127.0.0.1";
+        RETROARCH_LISTEN_PORT = 55355;
+        RETROARCH_READ_PACKET_TIMEOUT_MS = 64;
+        DELAY_MS_BETWEEN_READS = 5;
+        PROTOCOL_FRAMESKIP = -1;
+        MAPPER_VERSION = "0.0";
+        GITHUB_TOKEN = "";
+    }
+
     public AppSettings(IConfiguration configuration, ILogger<AppSettings> logger)
     {
         RETROARCH_LISTEN_IP_ADDRESS = configuration.GetRequiredValue("RETROARCH_LISTEN_IP_ADDRESS");
         RETROARCH_LISTEN_PORT = int.Parse(configuration.GetRequiredValue("RETROARCH_LISTEN_PORT"));
         RETROARCH_READ_PACKET_TIMEOUT_MS = int.Parse(configuration.GetRequiredValue("RETROARCH_READ_PACKET_TIMEOUT_MS"));
-        RETROARCH_DELAY_MS_BETWEEN_READS = int.Parse(configuration.GetRequiredValue("RETROARCH_DELAY_MS_BETWEEN_READS"));
-
-        RETROARCH_DELAY_MS_BETWEEN_READS = int.Parse(configuration.GetRequiredValue("RETROARCH_DELAY_MS_BETWEEN_READS"));
-
-        BIZHAWK_DELAY_MS_BETWEEN_READS = int.Parse(configuration.GetRequiredValue("BIZHAWK_DELAY_MS_BETWEEN_READS"));
+        DELAY_MS_BETWEEN_READS = int.Parse(configuration.GetRequiredValue("DELAY_MS_BETWEEN_READS"));
 
         PROTOCOL_FRAMESKIP = int.Parse(configuration["PROTOCOL_FRAMESKIP"] ?? "-1");
 
@@ -33,18 +43,47 @@ public class AppSettings
 
         GITHUB_TOKEN = configuration["GITHUB_TOKEN"] ?? "";
 
-        logger.LogInformation($"AppSettings initialized: RETROARCH_DELAY_MS_BETWEEN_READS: {RETROARCH_DELAY_MS_BETWEEN_READS} and BIZHAWK_DELAY_MS_BETWEEN_READS: {BIZHAWK_DELAY_MS_BETWEEN_READS}");
+        logger.LogInformation($"AppSettings initialized: DELAY_MS_BETWEEN_READS: {DELAY_MS_BETWEEN_READS}.");
     }
 
-    public string RETROARCH_LISTEN_IP_ADDRESS { get; }
-    public int RETROARCH_LISTEN_PORT { get; }
-    public int RETROARCH_READ_PACKET_TIMEOUT_MS { get; }
-    public int RETROARCH_DELAY_MS_BETWEEN_READS { get; }
+    [JsonPropertyName("RETROARCH_LISTEN_IP_ADDRESS")]
+    public string RETROARCH_LISTEN_IP_ADDRESS { get; set; }
 
-    public int BIZHAWK_DELAY_MS_BETWEEN_READS { get; }
+    [JsonPropertyName("RETROARCH_LISTEN_PORT")]
+    public int RETROARCH_LISTEN_PORT { get; set; }
 
-    public int PROTOCOL_FRAMESKIP { get; }
+    [JsonPropertyName("RETROARCH_READ_PACKET_TIMEOUT_MS")]
+    public int RETROARCH_READ_PACKET_TIMEOUT_MS { get; set; }
 
-    public string MAPPER_VERSION { get; }
-    public string GITHUB_TOKEN { get; }
+    [JsonPropertyName("DELAY_MS_BETWEEN_READS")]
+    public int DELAY_MS_BETWEEN_READS { get; set; }
+
+    [JsonPropertyName("PROTOCOL_FRAMESKIP")]
+    public int PROTOCOL_FRAMESKIP { get; set; }
+
+    [JsonPropertyName("MAPPER_VERSION")]
+    public string MAPPER_VERSION { get; set; }
+
+    [JsonPropertyName("GITHUB_TOKEN")]
+    public string GITHUB_TOKEN { get; set; }    
+
+    public string Urls => "http://localhost:8085";
+    public string AllowedHosts => "localhost";
+
+    public void Save()
+    {
+        var appSettingsPath = Path.Combine(Directory.GetCurrentDirectory(), "appsettings.json");
+        var settingsText = JsonSerializer.Serialize(
+            this, 
+            new JsonSerializerOptions()
+            {
+                WriteIndented = true,
+                TypeInfoResolver = MapperDtoContext.Default
+            }
+        );
+        File.WriteAllText(
+            appSettingsPath, 
+            settingsText
+        );
+    }
 }